Add text to the Create Volume dialog's Type menu
The first value in the "Type" dropdown menu of the Create Volume dialog
was empty. This causes some confusion to the user because there was not
enough information as to why the value was empty or what would happen
when creating a volume with the empty type.

Adding text "No volume type" to the first value of the dropdown menu
explains that there will be no volume type associated with the volume
at creation time.

Closes-Bug: #1325701
Change-Id: Iaff5d413d4abe9c59d48a679094ea37930abc4c1