Basic PoC for LB Vip <select> option label display
In "Manage Floating IP Associations": the port name for a Vip port is
"None" which is confusing for the user. If the Vip is attached to a load
balancer pool: it should be the name of the pool.

This commit demonstrates how the logic of FloatingIpManager.list_targets
could be updated to display the name of the LB attached to a Vip
potentially returned by port_list.