Fix keystone warnings while building docs
When building the documentation via run_tests.sh, doc/source/conf.py
explicitly sets the DJANGO_SETTINGS_MODULE environment variable,
overriding any existing value that it already has. Furthermore, it sets
it to point to a settings file that does not use keystone v3, which
is expected by the tests (and is the source of the warning).  Note that
when running unit tests via run_tests.sh, a different settings file is
supplied which DOES use keystone v3.

Change the doc/source/conf.py to only set DJANGO_SETTINGS_MODULE if it
is not already set.  Change tox.ini and run_tests.sh to set
DJANGO_SETTINGS_MODULE to the same settings file used by the unit tests,
with keystone v3 support.

Change-Id: Ib297e4188f2426cf575300998bc9d50f36e48f4f
Closes-Bug: #1257725