Change Admin Volumes to a tabgroup
Currently there is no Snapshots area in Admin Volumes. In order to
enable the reset snapshot state feature, We need a new tab for Snapshots
in the same manner as what's in Project Volumes. This submit is part 1
of the work, which is to refactor the Volumes panel and make it a
tabgroup. This is so that we can add another tab for Snapshots later on
(in part 2).

Note that these are only structural changes. There are no content
changes.

Partial-Implements: blueprint cinder-reset-snapshot-state
Change-Id: Ibadc4151e698dc6ea99009f6a63fc89c6d001946