Disable broken unit test (related to "Change Password")
The new django_openstack_auth 1.1.6 library surfaced a defect in this
test, whereby the test client does not properly log out the user. This
only affects the test, the functionality otherwise works fine.

This temporarily disables the test in order to repair the gate, while
waiting to replace it with a new unit test more reduced in scope and/or
an integration test.

Change-Id: I8779666c053a4835e22038a39a530fe5fc84b6a3
Closes-Bug: #1333144
(cherry picked from commit 5455b8f6f595633513b6521449ebc1985d5b9070)