Fix inconsistent inline help text in forms
Currently, help text is only shown for some input fields.
For example, if you go to 'Launch Instance', click on
the input field 'Instance Count' to see the help tooltip.
It is not displayed for checkboxes (BooleanFields) or dropdown menu,
even though help_text is provided in the code.

This change places a small 'info' icon beside each label (if help_text
is defined for that django form element) so that when you hover over it,
it will show you the corresponding help text.

This patch provides a uniform solution for all field and allows to user to
immediately see which fields have help.

Change-Id: Ie2d85503160071ba4062b3368e2ef79e7786343e
Closes-Bug: #1332738
Closes-Bug: #1330766