Allow table actions to be displayed in a menu
This adds the ability for table actions to be placed into an
actions menu above the table in addition to the current
button only format. This ability is not implemented in any
existing horizon table but is mainly meant for those who
build on horizon with tables that might have many actions.

To use a table actions menu, add a list of table action
classes to the table Meta class using the table_actions_menu
attribute. For example, to display the Soft Reboot and
Terminate actions in a menu on the project instances table
add this attribute to the project instances table Meta class:

    table_actions_menu = (SoftRebootInstance, TerminateInstance,)

Implements: blueprint table-action-menus
Change-Id: I02cd52f06be3fa11dd5fe0b45c8cc1e3bff5efdb