add httplib2 explicitly to requirements.txt
On Ubuntu 14.04+ whatever dark magic was installing httplib2 into
the venv for unit testing, is no longer doing it, as httplib2 was
not an explicit requirement.

It is clearly used in the code, and a dependency, so be explicit
about it.

Change-Id: I373b87dd4af414d58680d34c4de626ffa953cf57
(cherry picked from commit f712ed04ccb49e832142944d579e89fb7c14d394)