Add OS_REGION_NAME to openrc
The openrc file that is downloaded from Horizon is missing the
OS_REGION_NAME setting. This causes issues in environments that have
multiple regions. If the Region is none, ensure that it is unset after
sourcing the rc file to avoid and empty variable. This fix also cleans
up some language in the generated openrc file.

Change-Id: Ieab0b13afd79158c6bf603891164f01aff3aabf3
Closes-Bug: #1334791