Reinstate highlighting of selected container in Containers view
The AJAX updating of the containers is the culprit here in a couple of ways.

Firstly, the AJAX update (the short-circuit in MultiTableView.construct_tables)
doesn't have the context from the URL which indicates which of the rows should
be highlighted. This is because the current_item_id is only set on the
navigation table in the case where the whole browser is being rendered, rather
than just part of it. Shifting setting that attribute to all cases allows
lower-level access to the attribute.

Secondly, the handling of the AJAX update (in DataTable.maybe_preempt)
doesn't check the row it creates to see whether it should be selected.

Change-Id: Ibecb761a4bde25c1ea4e33a581dd0586ecc42ea0
Closes-bug: 1299828