Long container names no longer break the page
In the containers page, if the name of a container is too long, the
objects table is no longer visible and the table is out of the screen.
There's a screenshot in the bug itself.

Conflicts:
	openstack_dashboard/static/dashboard/less/horizon.less

Change-Id: I4b37d04fec462b2b2baad5a3c76db0d0f48e5387
Closes-Bug: 1314145
(cherry picked from commit d20bde3d7be621642faa300ba3ae28ab65c2c6b5)