Add a None option to the CONSOLE_TYPE setting
We where able to set the CONSOLE_TYPE in local_settings.py, but not able to
deactivate it.

We added a new option to CONSOLE_TYPE which is None.
To keep the legacy behaviour, CONSOLE_TYPE defaults to 'AUTO' and has to be
explicitly set to None to deactivate the console.

implements bp hide-vnc-console-access

Change-Id: Id23f3d5e37509b846858d480ad781f2b4ffc3590