Run router dashboard unit tests by default
Currently the "router" dashboard does not have its unit tests
run by default. This was beacuse the dashboard needed to be
turned on for these tests to be run and turning on this dashboard
was not mandatory.

However, now that the dashboards are pluggable, turning on the
router dashboard tests, and more specifically the Cisco N1K profile
tests is doable and necessary. This fix aims to remove the checks
put in place to run the tests only when the dashboard is available
and instead now run all the tests mandatorily

This fix was merged in J-2. Backporting to stable/icehouse
so that the 'router' dashboard and tests using profile
are run in Icehouse as well.
-----

cherry picked from commit 2f2312dfcccda1b421a80f27e6759d67f4df6ba1

Conflicts:

	openstack_dashboard/dashboards/project/instances/tests.py

    def test_launch_instance_post_with_profile(self):
        self.test_launch_instance_post(test_with_profile=True)

Note - Fixed 2 string messages in create_instance.py so as to
not affect translation and retain previous messages.
Additionally, fixed minor pep8 errors in instances/test.py
and networks/tests.py

Closes-bug: #1229819
Change-Id: I3c77a71db08ac611e5bdf9274af5a558bfda3327