Handle RequestURITooLong error in large instance table
Neutron API supports filtering in listing API. If long filter parameters
such as long UUID list are passed, it exceeds URL length limit and
RequestURITooLong is raised. In such case, we need to split filter
parameters into chunks.

This commit add a convenient method to handle the case, which split
a list of long filter parameter into chunks and repeat listing API call.
However, it is not easy to generalize this behavior because filtering
parameters are specific to caller's logic, so this patch only fixes
a problem in servers_update_addresses().

Change-Id: I63ec4bad722351f640797bcbec4a38d5972f135c
Closes-Bug: #1349841