Make project optional during User creation
For keystone V3, the default project is an optional field.
A Domain Admin account may only be interacting with Keystone
and does not need to be associated with a Project. Project
association is required for Domain Users that needs to use
services other than Keystone. Making this optional allows the
Cloud Admin to setup a Domain Admin account without the project
association.

Change-Id: I20dff330ce3fcbbb3be7e8512df66c01aa5e3458
Partial-Implements: blueprint domain-admin-support