Fix broken Subnet Wizard by updating to latest library
Updated jquery.bootstrap.wizard.js to the 2014/7/11 version.
Updated from https://github.com/VinceG/twitter-bootstrap-wizard. 
The bug was caused by changing bootstrap version without updating wizard.
This version "Supports Bootstrap 2.2.x, 2.3.x, 3.0".

(More specifically, the old wizard code was listening for a 'shown'
event that was never getting fired; the event is renamed 'shown.bs.tab'.)

Disconcertingly, there is no version number on the file, and the
comment at the top of the file says "1.0" - which is what it said
before. The timestamp is all we have to version it with.

Change-Id: I924df2d8ad20859688db261c2185c407450b174e
Closes-Bug: 1350733