TEMPLATE_DIRS must be a tuple
With Django 1.7, TEMPLATE_DIRS must be a tuple. The
definition in horizon/test/settings.py must therefore
have a leading comma. Note that other definitions of
TEMPLATE_DIRS are already like this in Horizon, so
this just fixes the one instance which is wrong.

Note that it is required to support Django 1.7 in
Icehouse, because we're planning to release Debian
Jessie with both Django 1.7 and Icehouse. This is to
*support* Django 1.7, not at all to *switch* to it,
which is very different.

Change-Id: I4d9aa9e716e4cf6c2b56fd25ab8fde9187fc5108
(cherry picked from commit daf9e5c772f82b783ce2255f7cd347ef397efcaf)