Rename add_error methods: Django 1.7 conflict
In Django 1.7, Forms have an add_error() method and our definition
was conflicting with Django's method. Thanks to hertzog@debian.org
for the patch.

Note that it is required to support Django 1.7 in Icehouse, because
we're planning to release Debian Jessie with both Django 1.7 and
Icehouse. This is to *support* Django 1.7, not at all to *switch* to
it, which is very different.

Close-bug: #1352919
Change-Id: I1f558cf7f3198a68629738a011e9cae0c3749320
(cherry picked from commit 04dd1568c1c927901085941124e4c6852183f4f0)