template to rely on the the "id" attribute
The exception fixed is the following:
NoReverseMatch: Reverse for 'detail' with arguments '('',)' and keyword
arguments '{}' not found. 1 pattern(s) tried:
[u'project/volumes/(?P<volume_id>[^/]+)/$']

The test data doesn't include a volumeId attribute to volumes but they
contain an id attribute.

Here I modify the template to rely on the the "id" attribute but maybe
real data has a volumeId attribute and it's the test data in
openstack_dashboard/test/test_data/nova_data.py that needs to be
modified instead.

Note that it is required to support Django 1.7 in Icehouse, because
we're planning to release Debian Jessie with both Django 1.7 and
Icehouse. This is to *support* Django 1.7, not at all to *switch* to
it, which is very different.

Change-Id: Ia561c2db1c59e51fb669e1ad6636d9b3bc9f7676
(cherry picked from commit 149f842017f44b24f83a904498a4df6c6d2b6cc9)