Add workaround for StaleElementReferenceException
StaleElementReferenceException occurres when accessing web element that
was reloaded but looks the same. It happens for example when trying to access
form under Project/Data processing/image registry -> register image. It
is ussually caused by javascript that is loading the page or by mistake
in script logic.

This patch wraps the WebDriver and WebElement classes and catches the
occurrence of StaleElementReferenceException and tries to find the web
element again. Till maximum limit is reached or top element is reached
as this suggest that mistake was made by programmer.

Closes-Bug: #1353342

Partially implements blueprint: selenium-integration-testing

Change-Id: I8e24e952e0fb81bed21b7139870d6d7506509c93