Integration tests - switch_window method
I added the switch_window method to pageobject, so that it can be
used by all pages.
The method switches focus between the webdriver windows.
Args:
- window_name: The name of the window to switch to.
- window_index: The index of the window handle to switch to.
If the method is called without arguments it switches to the
last window in the driver window_handles list.
In case only one window exists nothing effectively happens.
Usage:
page.switch_window('_new')
page.switch_window(2)
page.switch_window()

The motivation of writing it was to prevent the inconsistencies
encountered lately, and the common use of a deprecated similar method.
It is not meant for handling pop-ups.

I also modified test_dashboard_help_redirectionת so it now uses
the switch_window method.

Partially implements blueprint: selenium-integration-testing
Closes-Bug: #1354824

Change-Id: Ic90accc37af292994656e1b65ad5c8527b50aaed