Ability to change modal form backdrop element behavior
By default modal form is closed after clicking outside of it on a dark
backdrop (data-backdrop="true"). For the forms with a lot of fields
this could lead to a loss of data the user had entered before he has
unintentionally clicked outside of the form. So more preferable option
for such forms (and workflows/client-side modals) would be
data-backdrop="static".

DocImpact
Change-Id: I56fca9131a4e4d3991d53184ee1332e84138cdc2
Partially-Implements: blueprint form-template-to-view
Related-Bug: #1062065