Edit image - made format field as editable
While editing image via CLI we can edit/change
the disk format via '--disk-format' option.
In horizon webpage, edit image form, the format
field is not allowing to edit the disk format since
it is set as non-editable field.
It is differ from the CLI functionality so changed
the format field is editable to sync with CLI.
the choice field is displayed with available choices,
the user can select one from the list.
Change-Id: I6bea35ff8e9d9ae2d170969b9a7fce1b2d657d09
Closes-Bug: #1325123