Escape RegEx characters from table quick search text
This fix escapes any regular expression characters from the table
quick search text so that the filter works as the user would
expect.

A new string utility module is created for the legacy javascript
string utils and this new function for escaping regex chars is
placed there along with the existing function for escaping HTML
chars.

Closes-Bug: #1327276
Change-Id: I514bbe7ae9e1f319ced693e9a3da2881a86d4f57