Fix IE spacing for pie charts
The svg pie charts are being mis-aligned in IE because ID doesnt
correctly process the height and width attributes on an svg element
Using a viewBox resolves the issue and doesnt appear to cause any
side effects in other browsers. Note that the IE charts will scale
slightly as the browser window changes. Chrome and FF do not
exhibit this because they strictly apply the height and width

Change-Id: Id45b1f11dd964425926795f6d07ab7bdc361f6ee
Closes-bug: 1342125