Document token hash algorithm option
With https://review.openstack.org/#/c/116509/ ,
django-openstack-auth will support a new option for the token hash
algorithm. This adds the documentation to Horizon's local settings
example file.

This is for security hardening. The token hash algorithm defaults
to MD5, which is considered too weak due to the potential for hash
collisions. Some security standards require a SHA2 hash algorithm to
be used.

DocImpact
SecurityImpact

Change-Id: I6774b9b7215d191259586e4721e357487bb777cd
Closes-Bug: #1174499