Fix dynamic select layout when help block is displayed
This fixes the layout of the dynamic select widget when the help
block is displayed. To reproduce this error you must be using the
keystone v2 API so that the Primary Project field will be required
when creating a user. If you leave it blank when submitting the
form the help block will be displayed.

Closes-Bug: #1361725
Change-Id: Ied4bb5d22b7e8a1948330503516cba2aa016417f