Add 'Start/Stop Instance' to Instances table actions
Some users have been reporting that they would like the ability to
start multiple instances at once, for instance after an upgrade it's
common for the instances to be shut off and need to be manually
restarted.

This patch adds 'Start Instances', 'Stop Instance' and moves
'Soft Reboot Instances' to the Instances table actions menu which is
a recently implemented feature.

Change-Id: I3c3b29a92fe61e6330b60c3d2228d899c409c23a
Closes-bug: #1354395