Remove and deprecate dashboards config from settings file
Dashboards are now configured via the pluggable dashboard/panel
mechanism and the order can be specified in the pluggable config.
On the other hand, we have HORIZON "dashboards" and its tuple order
overrides the order specified in the pluggable config.
It is confusing to be able to configure the order of dashboards
in two places, and we no longer need to have "dashboards" config
parameter in settings.py and a sample local_settings.py.

* If HORIZON_CONFIG contains "dashboards" and update_dashboards
  is called, a deprecation warning will be displayed. It is useful
  to warn operators to change OpenStack Dashboard settings.

* HORIZON_CONFIG "dashboard" parameter itself still works and
  there is no chnage in the behavior, so it is a safe change
  and have no upgrade impact.

* This change also removes default_dashboard from settings because
  it can be configured via pluggable dashboard config too.

After removing "dashboards" config, the order of the dashboards
are now controlled by the order of openstack_dashboard/enabled and
"Identity" dashboard is displayed before "Settings" dashboard
in the navigation menu.

Closes-Bug: #1366270

Change-Id: Ifa35128691c4fa4843b3c3c051682392e59905a6