Determine environment version based requirement files
It is painful to increment the venv version manually because
requirements.txt and test-requirements.txt are now updated by cron job.
This commit changes run_tests.sh to just store two requirements
files to .venv/environments file and check if it is up-to-date.
(Note that the environment cache file is moved to .venv directory
as there is no reason to store it in the top directory.)

Closes-Bug: #1376485
Change-Id: Ie44ccf6e2e65890baca3f316468254b600c518b1