Create Subnet button should respect quotas
The create subnet button (visible for each network in the networks list and on
the detail page when looking at an individual network) should be disabled if
the quota for subnets is at capacity.

Tests have been added to check that when the quota is exceeded the buttons are
disabled.

Change-Id: I7bba89c2d89a39220900480f0091bdc6215505fe
Closes-bug: 1369779