Dismiss flavor detail pop-over on users next click
In the Instances table under Projects, when we click the flavor
link under Size column, a flavor details pop-over shows up.
Currently we cannot dismiss this pop-over unless user clicks the
flavor link again.

This patch is intended to fix this issue by dismissing the pop-over
on the user's next click.

Change-Id: I46e4d368ef174f04f245be8eef8b3b244c5d1201
Closes-Bug: #1370133