Handle correctly protocol 'ANY' in firewall rule
Currently horizon allows to create firewall rule with protocol == 'ANY'
but horizon crashs if you try to update such rule because UpdateRule
assumes that protocol is a string and fails to convert it to uppercase
if protocol is ANY/None. Moreover horizon does not allow to update
protocol to 'ANY' despite neutron supports it.

This change updates UpdateRule in order to convert correctly ANY to
None and vice versa. It allows also to update firewall rule protocol to
ANY.

Closes-Bug: #1324382
Closes-Bug: #1370485
Change-Id: I5748b8e6c1cf0ef0d7a4317016492e064b5e8918