Made Admin State choice more human readable
The Admin State choices in create forms throughout horizon
were not very clear, this patch presents that choice in a much more
readable and understandable way. The field is changed from a boolean
field to a choice field with the different states, making it more
obvious what is being decided.

Change-Id: I7091e5ccf91c57d3f64e2d4415bed952d09ada71
Closes-Bug: 1370322