Make Image Description an input field instead of a textarea
The glanceclient passes the value of the "Description" field as
HTTP headers to the API, completely unmodified.  If there are
newlines in the content, it corrupts the headers and creates havoc.

Encoding/decoding the results is not viable, since arbitrary (and
not necessarily controlled by anyone) clients can retrieve the data.  So as
a dodge, use a normal <input> field for Description instead of
<textarea> field, as newlines are not permitted in the former.

Closes-Bug: 1370732
Change-Id: I0231eaa693e7f36699fda6e011db8c976b639411