Move datepicker language js inclusion outside compress
The datapicker language support works by including a different
js file based on the language of the user.  This will not
work if the js is included at compress time becuase only 1
language can be included.

Additionally cleaning up the logic in case the locale
isn't set (as is the case at compression time) to prevent
an improper path from being build.

Change-Id: I953dde0e7ee4a876376faf28c6f0bb859a3dc6b6
Closes-Bug: #1373430