Do not override user-defined Device Size
This change introduces a flag which identifies whether or not the
user has manually set the Device Size field when launching an
instance with a new volume. If the user has made a change then the
value will not be overridden by the logic that sets the Device Size
based on the selected image/flavor unless the user-defined value is
too small.

Change-Id: I1f02fd386a9b6836fb95433d9cb819af4f267d6c
Closes-bug: 1370080