Fixes resource name problem in "Resources Usage" tab
This patch replaces the resource id with resource name in the
""Resources Usage" Tab when selecting the "Group By" Filter to
"--".
In this fix i has created a mapping between the meter_name and the
api_type and upon depending the meter_name, the code make the
appropriate api call by passing the resource_id.

Closes-bug: #1271342

Change-Id: Ie851cb34c9f16c47ddd7eed4a19c77398e350048