Enable launching instance from zero-size image snapshot
Remove the general hardcoded minimum of 1GB volume size on the create
instance workflow, and selectively enforce it when the volume size field
is used and editable: namely, when launching an instance from an image
and creating a new volume.

This general minimum was preventing the launch of instances from image
snapshots (whose size is reported as 0) by causing browser errors as the
browser tried to highlight the error of the invisible size field.

Closes-Bug: 1374931
Change-Id: Ied0eb41f5198d935bbb0896e5004ea78176bc2b8