Allow to hide tab title bar when only one tab is available
In detail pages with a single tab, tab title bars are redundant
and unnecessary. For such pages, it would be nice if a tab bar
is hidden.

This commit introduces show_single_tab attribute to TabGroup
class. The default value is set to False because hiding a tab bar
for a single tab page looks appropriate in most cases.

Tab mechanism makes developers to add new tabs for extensibility.
If we remove tab view and change such view into a normal DetailView
we will lose the extensibility. After discussion with some Horizon
developers, it sounds nice to hide tab bar for a single tab page.

Closes-Bug: #1382546
Change-Id: I90276c531d0a017e3970b48290ae730ea3f1ecad