Invalid instance_snapshot causes KeyError
When creating a nova instance, if the instance_snapshot_id passed in is
invalid, a KeyError exception results.  This likely happens when another
window or perhaps a command line tool deletes the snapshot before the
page is posted back.  Memoize decorator fix fixes an error which masks
the error listed above, enabling better debugging.

Closes-Bug: 1386826
Change-Id: I195b4df84da3bd360b155db9ae733c34575ebe7e