Integration tests - 3 new form fields
I added 3 new form fields, that are frequently used
in different horizon's forms:
* CheckBoxFormField - with the options to mark, unmark
and check if already marked.
* ChooseFileField - with a choose file action.
* TextAreaFormField - intended for multi-line text input box.

The reason I separated this change from the Images page patch
is to allow others to use it in their patches without forcing
the images as a dependency.

Partially implements blueprint: selenium-integration-testing

Change-Id: I94260db6089e44e66bf19b0ede172eec1756cdc5