Remove duplicated headings from horizon tables
Many panels in horizon only contain a single table, in these cases the
tables often have the same name as the panel, leading to duplicated
headings, e.g. Instances (panel heading) then Instances (table
heading). This patch adds an attribute the table to hide or show the
title, since the majority of panels have this duplication this flag
is set to hide the heading by default allowing a developer to re-enable
it for specific cases.

Tabbed panels also often lead to duplicated titles between the panel
heading and the tabs themselves, that is not fixed, however how this
patch reduces the duplication in the case of Volumes (panel name),
Volumes (tab name) and Volumes (table name), by hiding the table
heading.

Change-Id: I65efa53cc881417288e5c5b29bfa7d539e97c7dd
Closes-Bug: 1379044