Moving policy engine implementation
This is the first step in moving the policy engine to
django_openstack_auth. It makes the policy check method pluggable
in openstack_dashboard as it is in horizon. To do this, the wrapper
around the policy engine is moved to a new file isolated from the
policy check method. A thin check method is left in policy.py to
allow the code and policy mix-ins to behave as now. The existing
policy test file is also moved for relocation. A new test file has
been added to exercise the simpler check method.

Once the actual engine is moved to django_openstack_auth, we'll be
able to remove policy_backend.py and the test version of
policy_backend.py as well.

Partially Implemenents: blueprint move-policy-engine

Change-Id: I591bcbe69bf255f1aa0346fdf869fac86d2e6d3d