Create a reusable angular-based wizard
Motivation:
The initial driving motivation is to redesign the launch instance wizard
based on a modern, responsive design that is based on client side
rendering using angular.

Description:
The wizard will be as lightweight as possible, but still allow for
customizable text, localization, customizable per-worfklow number of
steps.

Design:
* Angular-based.
* Using Angular Bootstrap UI modal widget.

Co-Authored-By: Shaoquan Chen <sean.chen2@hp.com>
Co-Authored-By: Szymon Wroblewski <szymon.wroblewski@intel.com>
Implements: blueprint initial-reusable-angular-wizard
Change-Id: I6d7530b6723764d3ca117e16eeae78c0ab35e48e