Cleanup angular 1.2.1.1 dependencies
Adding version 1.3.7

1.3.x angular is needed for Horizon launch instance work.

This was dependent on the following (done):
https://review.openstack.org/#/c/142520/
Xstatic-Angular 1.3.7 includes the following packages:
XStatic-Angular-Cookies
XStatic-Angular-Mock

They can be removed, but this apparently needs to be done
in a multi-step process. The interdependency chain,
(horizon has dependency on the old packages) causes
an integration test to fail. So, if we follow the normal
process, then we have to do it in the following order.
1) The newer pakage has to be added to global requirements
2) a) The new dependency has to be added to Horizon
   b) Old dependencies must be removed from Horizon
3) The older packages must be removed from global requirements
   - change 145096

Note: This patch includes the 1.3.7 dependency update, but
that will be handled by the auto update when global requirements
is merged and this should rebase automatically on top of that.

Partially Implements: blueprint launch-instance-redesign

Change-Id: I83770224dd6b8df2da4e5cb2efaa3ec932b6e4a1