security group not mandatory to create instance
as per the nova api, the security group is not
mandatory option to create an instance.

but in the launch instance form, it is marked as
mandatory and it is not allowed to create/launch
an instance if no security group was selected.

removing the mandatory field mark '*' for security
group in the launch instance form.

Change-Id: I80d61665a0b794bc1b354f47834e79cb33cbacd1
Partial-Bug: #1405109