Set proper z-index value on messages and navbar
The navbar should have been using the built in navbar-fixed-top class
as it is a fixed top navbar.  After added the correct class, the
correct z-index value was added via Bootstrap, then the messages
container simply needed to add a z-index that would appropriately
place it above dropdowns.

The user menu was simply a symptom of the overall problem, where all
the dropdowns in the navbar needed to behave the same way.

The z-index variables of the theme were used, therefore the
incorrect stacking of the messages over the spinner and the modal
were fixed at the same time.

It was noted that the right padding of the messages was using the
incorrect padding variable.  This was corrected.

Closes-bug: #1408640
Co-Authored-By: Diana Whitten <hurgleburgler@gmail.com>
Change-Id: I1d59049d43e74c2a897673307593993f4291da39