Fix disappearing Sahara workflow navbar in 'Jobs' panel
The jquery selector used before in Sahara-specific js was too wide so
it could seize even the modal workflows from the depth of modal stack
(i.e. currently hidden workflows) - which lead to damaging these
workflows' DOM - the fact which became evident during unwinding the
modal stack. The patch contains the fix for the 'Jobs' panel as well
as for other copy-pasted snippets of workflow-handling code.

Change-Id: Id2544c18c27ae3432e6fa43f19cd1459ffb34871
Closes-Bug: #1414586