Prevent flash of left nav panels when loading page
In the left nav, only the panel group containing the panel for the current page
should be expanded. However, in the Chrome browser, all other panels in the
same dashboard are momentarily shown before being hidden. The flash of content
happens in the time between the page's template rendering the panels and the
JavaScript being run to hide the non-active panels. Prevent the flash by hiding
non-active panels in the template instead of using JavaScript to hide the
panels later.

Closes-Bug: 1408513
Change-Id: I8fde47f4a0d6b7415998cf84fa4974bbe39728eb