Update Metadata Widget to angular 1.3.7
Following stopped working with update to angular 1.3.7

 - Available metadata (type a filter to see it)
 - Adding metadata from left to right
 - Existing metadata
 - Saving metadata

There are numerous things that have changed in angular 1.3.7
from the previous angular 1.2.x.

The view of available metadata and existing metadata was affected
by the array filter no longer handling undefined for the $ input argument.

The + over and saving metadata were affected by this change:
!! in filters is now !null
 http://stackoverflow.com/questions/18644412/angularjs-filter-not-null

ng-pattern changed in 1.3.x

http://stackoverflow.com/questions/26439161/
angularjs-validation-behaves-differently-for-1-3-0-and-1-2-6

It now works in flavors and host aggregates, but still has an issue in images.
The images issues appears to be unrelated, but caused by a javascript error
on the images page itself.

Change-Id: I4a2c9f29fdbc6bfc316e3b3b0000817d1516aaa2
Partial-Bug: #1410583