equal(=) added in project filter option
In admin's instance table, the filter with
'project' option filtering by exact matching
not by partial match.

the equal(=) symbol is missing at the end of the option.
there should be the = symbol to tell the user
it is doing exact match like other options.

in this patch the = symbol is added at the end of
the project option.

Change-Id: I2f81e4168d33d2dd34963039db85456a17cdbaf3
Closes-Bug: #1413534