Glance Rest API for Angular Front End
This is the glance direct / non-abstracted rest API
for the angular front end.

This initial patch will only implement the APIs needed to
support the launch instance work.

This was refactored out of change: 141273
All v1 / v2 work is being done separately in change: 150084

Based on the work done for Keystone here:
https://review.openstack.org/#/c/150636

Partially Implements: blueprint launch-instance-redesign
Co-Authored-By: Richard Jones <r1chardj0n3s@gmail.com>
Change-Id: I28735e242d610235695b1070a15f739e0d50b696