Read access to config via REST
We have a lot of new angular widgets and panels coming, and some of them
will need access to configs stored in settings. Right now, the default way to
get them is through _conf.html or _script.html. We should really fetch them
through the new REST layer instead. This ensures we are following the same
pattern already set up for angular panels and services.

This version:
- add localstorage for caching config front-end

Change-Id: I94ce8ef3e8367be99bbba2f6ddd7e4529d819d58
Co-Authored-By: Richard Jones <r1chardj0n3s@gmail.com>
Implements: blueprint config-restful