Launch Instance Model
This is the M part in MVC for launch instance that will use the APIs
to load data and send request to server via the APIs.

This model is injected into to LaunchInstanceCtrl and available to all
controllers for every step in the work flow.

To access:
 - Add to local_setting.py:
     LAUNCH_INSTANCE_NG_ENABLED = True
 - Then go to:
     http://localhost:9000/project/instances/
 - Then click the left "Launch Instance" button.

To disable the legacy launch instance button:
 - Add to local_setting.py:
    LAUNCH_INSTANCE_LEGACY_ENABLED = True

Partially Implements: blueprint launch-instance-redesign
Co-Authored-By: Travis Tripp <travis.tripp@hp.com>
Co-Authored-By: Neill Cox <neill.cox@ingenious.com.au>
Co-Authored-By: Richard Jones <r1chardj0n3s@gmail.com>
Co-Authored-By: Tyr Johanson <tyr@hp.com>
Co-Authored-By: Matt Borland <matt.borland@hp.com>
Change-Id: I040f13f7d1fd336fcd1070bdbe0c81bd5a828d43