cancel next previous buttons are aligned properly
In create network form, the buttons in the footer are
not aligned properly, the image is attached in the bug report.
Each buttons are inserted in separate div blocks, this is causing
the misaligned view. now everything is put it in a single div and
it is aligned in a single row.
Change-Id: I6df8524865483dc8ea259f555a56a4a773dabdbf
Closes-Bug: #1379931